+++ WalletWasabi.Fluent/Helpers/CoinHelpers.cs
@@ -1,6 +1,8 @@
 using WalletWasabi.Blockchain.Analysis.Clustering;
 using WalletWasabi.Blockchain.TransactionOutputs;
+using WalletWasabi.Fluent.Models;
 using WalletWasabi.Models;
+using WalletWasabi.Wallets;
 
 namespace WalletWasabi.Fluent.Helpers;
 
@@ -32,4 +34,21 @@ public static class CoinHelpers
 	}
 
 	public static int GetConfirmations(this SmartCoin coin) => coin.Height.Type == HeightType.Chain ? (int)Services.BitcoinStore.SmartHeaderChain.TipHeight - coin.Height.Value + 1 : 0;
+
+	public static PrivacyLevel GetPrivacyLevel(this SmartCoin coin, Wallet wallet)
+	{
+		var anonScoreTarget = wallet.AnonScoreTarget;
+
+		if (coin.IsPrivate(anonScoreTarget))
+		{
+			return PrivacyLevel.Private;
+		}
+
+		if (coin.IsSemiPrivate())
+		{
+			return PrivacyLevel.SemiPrivate;
+		}
+
+		return PrivacyLevel.NonPrivate;
+	}
 }
