+++ WalletWasabi.Fluent/Models/PrivacyLevel.cs
@@ -1,0 +1,9 @@
+namespace WalletWasabi.Fluent.Models;
+
+public enum PrivacyLevel
+{
+	None,
+	SemiPrivate,
+	Private,
+	NonPrivate
+}
