+++ WalletWasabi.Fluent/ViewModels/Wallets/Send/PrivacySuggestions/LoadingSuggestionViewModel.cs
