+++ WalletWasabi.Fluent/Views/Wallets/Send/CoinjoinMoreSuggestionControl.axaml
@@ -1,0 +1,20 @@
+<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+             x:Class="WalletWasabi.Fluent.Views.Wallets.Send.CoinjoinMoreSuggestionControl">
+  <StackPanel Spacing="20">
+    <Panel MinHeight="80" MinWidth="100">
+      <StackPanel Spacing="10" VerticalAlignment="Top">
+        <PathIcon Data="{StaticResource wallet_action_coinjoin}"
+                  Foreground="{StaticResource SystemAccentColor}"
+                  DockPanel.Dock="Top" />
+
+        <TextBlock TextWrapping="Wrap" Text="Improve your privacy" TextAlignment="Center" />
+
+        <TextBlock Text="and retry your transaction later" TextAlignment="Center" FontSize="12" Opacity="0.6" Margin="0 30 0 0" />
+      </StackPanel>
+    </Panel>
+  </StackPanel>
+</UserControl>
