+++ WalletWasabi.Fluent/Views/Wallets/Send/CoinjoinMoreSuggestionControl.axaml.cs
@@ -1,0 +1,17 @@
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace WalletWasabi.Fluent.Views.Wallets.Send;
+
+public partial class CoinjoinMoreSuggestionControl : UserControl
+{
+	public CoinjoinMoreSuggestionControl()
+	{
+		InitializeComponent();
+	}
+
+	private void InitializeComponent()
+	{
+		AvaloniaXamlLoader.Load(this);
+	}
+}
