+++ WalletWasabi.Fluent/Views/Wallets/Send/PrivacySuggestions.axaml
@@ -1,0 +1,138 @@
+<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:c="using:WalletWasabi.Fluent.Controls"
+             xmlns:i="using:Avalonia.Xaml.Interactivity"
+             xmlns:vm="using:WalletWasabi.Fluent.ViewModels.Wallets.Send"
+             xmlns:behaviors="clr-namespace:WalletWasabi.Fluent.Behaviors"
+             xmlns:send="clr-namespace:WalletWasabi.Fluent.Views.Wallets.Send"
+             x:DataType="vm:TransactionPreviewViewModel"
+             x:Class="WalletWasabi.Fluent.Views.Wallets.Send.PrivacySuggestions">
+  <Panel>
+    <Border Cursor="Hand" Padding="10 0 10 10" IsVisible="{Binding !!PrivacySuggestions.Suggestions.Count}">
+      <Border.Styles>
+        <Style Selector="FlyoutPresenter">
+          <Setter Property="MaxWidth" Value="1000" />
+        </Style>
+      </Border.Styles>
+      <Interaction.Behaviors>
+        <behaviors:BindableFlyoutOpenBehavior IsOpen="{Binding PrivacySuggestions.IsOpen}" />
+      </Interaction.Behaviors>
+      <FlyoutBase.AttachedFlyout>
+        <Flyout Placement="Bottom" ShowMode="TransientWithDismissOnPointerMoveAway"
+                IsOpen="{Binding PrivacySuggestions.IsOpen, Mode=OneWayToSource}">
+          <Panel MaxWidth="540">
+            <Panel.Styles>
+              <Style Selector="ListBoxItem">
+                <Setter Property="Cursor" Value="Hand" />
+              </Style>
+              <Style Selector="ListBoxItem /template/ ContentPresenter#PART_ContentPresenter">
+                <Setter Property="Background" Value="Transparent" />
+              </Style>
+
+              <Style
+                Selector="ListBoxItem:selected /template/ ContentPresenter#PART_ContentPresenter, ListBoxItem:pointerover /template/ ContentPresenter#PART_ContentPresenter">
+                <Setter Property="Background" Value="Transparent" />
+              </Style>
+
+              <Style Selector="ListBoxItem Border#PART_MainContentBorder">
+                <Setter Property="Padding" Value="30,25,30,25" />
+                <Setter Property="CornerRadius" Value="4" />
+                <Setter Property="MaxWidth" Value="250" />
+                <Setter Property="MaxHeight" Value="300" />
+              </Style>
+
+              <Style Selector="ListBoxItem Border#PART_DecorationBorderUnselected">
+                <Setter Property="Padding" Value="15 25 15 25" />
+                <Setter Property="CornerRadius" Value="4" />
+                <Setter Property="Margin" Value="5" />
+
+                <Setter Property="BorderBrush" Value="#14FFFFFF" />
+                <Setter Property="BorderThickness" Value="1" />
+                <Setter Property="Background" Value="{StaticResource OptimizePrivacyOptionSelectedColor}" />
+                <Setter Property="BoxShadow" Value="{StaticResource OptimizePrivacyOptionBoxShadow1}" />
+              </Style>
+
+              <Style Selector="ListBoxItem Border#PART_DecorationBorderHover">
+                <Setter Property="Padding" Value="15 25 15 25" />
+                <Setter Property="CornerRadius" Value="4" />
+                <Setter Property="Margin" Value="5" />
+
+                <Setter Property="BorderBrush" Value="#2AFFFFFF" />
+                <Setter Property="BorderThickness" Value="1" />
+                <Setter Property="Background" Value="{StaticResource OptimizePrivacyOptionSelectedColor}" />
+                <Setter Property="BoxShadow" Value="{StaticResource OptimizePrivacyOptionBoxShadow2}" />
+
+                <Setter Property="Opacity" Value="0" />
+              </Style>
+
+              <Style Selector="ListBoxItem:pointerover Border#PART_DecorationBorderHover">
+                <Setter Property="Opacity" Value="1" />
+              </Style>
+            </Panel.Styles>
+
+            <DockPanel>
+              <TextBlock Text="Suggestions to improve your privacy" HorizontalAlignment="Center" DockPanel.Dock="Top" />
+
+              <ListBox DockPanel.Dock="Top" Background="Transparent" VerticalAlignment="Center"
+                       HorizontalAlignment="Center"
+                       Items="{Binding PrivacySuggestions.Suggestions}" SelectedItem="{Binding PrivacySuggestions.SelectedSuggestion}">
+                <i:Interaction.Behaviors>
+                  <behaviors:ListBoxPreviewBehavior PreviewItem="{Binding PrivacySuggestions.PreviewSuggestion, Mode=TwoWay}" />
+                </i:Interaction.Behaviors>
+                <ListBox.Styles>
+                </ListBox.Styles>
+                <ListBox.ItemsPanel>
+                  <ItemsPanelTemplate>
+                    <WrapPanel Orientation="Horizontal" ItemWidth="270" ItemHeight="230" />
+                  </ItemsPanelTemplate>
+                </ListBox.ItemsPanel>
+                <ListBox.DataTemplates>
+                  <DataTemplate DataType="vm:SendOnlySafeCoinsSuggestionViewModel">
+                    <c:SuggestionItem>
+                      <send:SendOnlySafeCoinsSuggestionControl />
+                    </c:SuggestionItem>
+                  </DataTemplate>
+                  <DataTemplate DataType="vm:ChangeAvoidanceSuggestionViewModel">
+                    <c:SuggestionItem>
+                      <send:ChangeAvoidanceSuggestionControl />
+                    </c:SuggestionItem>
+                  </DataTemplate>
+                  <DataTemplate DataType="vm:CoinjoinMoreSuggestionViewModel">
+                    <c:SuggestionItem>
+                      <send:CoinjoinMoreSuggestionControl />
+                    </c:SuggestionItem>
+                  </DataTemplate>
+                </ListBox.DataTemplates>
+              </ListBox>
+
+              <Button DockPanel.Dock="Top" Command="{Binding ShowOtherSuggestionsCommand}"
+                      Content="More options..."
+                      Classes="h8 plain activeHyperLink" HorizontalAlignment="Center"
+                      IsVisible="{Binding PrivacySuggestions.CanShowOtherSuggestions}" />
+            </DockPanel>
+          </Panel>
+        </Flyout>
+      </FlyoutBase.AttachedFlyout>
+      <Viewbox Height="20">
+        <PathIcon Data="{StaticResource shield_regular}" Foreground="{DynamicResource WarningMessageForeground}" />
+      </Viewbox>
+    </Border>
+
+    <Border Padding="10 0 10 10" Cursor="Hand" IsVisible="{Binding CurrentTransactionSummary.MaxPrivacy}">
+      <Interaction.Behaviors>
+        <behaviors:ShowFlyoutOnPointerOverBehavior />
+      </Interaction.Behaviors>
+      <FlyoutBase.AttachedFlyout>
+        <Flyout Placement="Bottom" ShowMode="TransientWithDismissOnPointerMoveAway">
+          <DockPanel>
+            <TextBlock Text="Your privacy is well protected in this transaction" DockPanel.Dock="Top" />
+            <Image Margin="0 20" Source="{StaticResource privacy_indicator_good}" Height="100" Width="100" />
+          </DockPanel>
+        </Flyout>
+      </FlyoutBase.AttachedFlyout>
+      <Viewbox Height="20">
+        <PathIcon Data="{StaticResource shield_regular}" Foreground="{DynamicResource PrivacyLevelStrongBrush}" />
+      </Viewbox>
+    </Border>
+  </Panel>
+</UserControl>
