+++ WalletWasabi.Fluent/Views/Wallets/Send/PrivacySuggestions.axaml.cs
@@ -1,0 +1,17 @@
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace WalletWasabi.Fluent.Views.Wallets.Send;
+
+public partial class PrivacySuggestions : UserControl
+{
+	public PrivacySuggestions()
+	{
+		InitializeComponent();
+	}
+
+	private void InitializeComponent()
+	{
+		AvaloniaXamlLoader.Load(this);
+	}
+}
