+++ WalletWasabi.Fluent/Views/Wallets/Send/SendOnlySafeCoinsSuggestionControl.axaml
@@ -1,0 +1,24 @@
+<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
+             x:Class="WalletWasabi.Fluent.Views.Wallets.Send.SendOnlySafeCoinsSuggestionControl">
+  <StackPanel Spacing="20">
+    <Panel MinHeight="80" MinWidth="100">
+      <StackPanel Spacing="10" VerticalAlignment="Top">
+        <PathIcon Data="{StaticResource money_24_regular}"
+                  Foreground="{StaticResource SystemAccentColor}"
+                  DockPanel.Dock="Top" />
+
+        <TextBlock TextWrapping="Wrap" Text="Send private funds only:" TextAlignment="Center" />
+        <StackPanel Spacing="10">
+          <TextBlock Text="{Binding Amount}" FontSize="20" HorizontalAlignment="Center"
+                     Foreground="{DynamicResource SystemAccentColor}" />
+          <TextBlock Text="{Binding AmountFiat}" FontSize="12" HorizontalAlignment="Center" Opacity="0.6" />
+          <TextBlock Text="{Binding DifferenceFiat}" FontSize="12" HorizontalAlignment="Center" Opacity="0.6" />
+        </StackPanel>
+      </StackPanel>
+    </Panel>
+  </StackPanel>
+</UserControl>
