+++ WalletWasabi.Fluent/Views/Wallets/Send/SendOnlySafeCoinsSuggestionControl.axaml.cs
@@ -1,0 +1,17 @@
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace WalletWasabi.Fluent.Views.Wallets.Send;
+
+public partial class SendOnlySafeCoinsSuggestionControl : UserControl
+{
+	public SendOnlySafeCoinsSuggestionControl()
+	{
+		InitializeComponent();
+	}
+
+	private void InitializeComponent()
+	{
+		AvaloniaXamlLoader.Load(this);
+	}
+}
